refactor: extract transform + mouse-global free funcs from TrussC.h; tcNode.h declares its deps
Pull the matrix / style-stack transform wrappers (pushMatrix / translate /
rotate / scale / ...) into tc/graphics/tcTransform.h, and the window-space mouse
state + getters (internal::mouseX/.., getGlobalMouseX/..) into
tc/app/tcMouseGlobal.h, so lower-level headers can depend on them directly
rather than on definitions buried in the TrussC.h body.

tcNode.h now includes its direct dependencies in dependency order and no longer
self-includes TrussC.h. The self-include was a no-op-or-trap: it could not make
tcNode.h standalone (the re-include is a #pragma once no-op while TrussC.h is
mid-flight) and only masked the rule that apps must include <TrussC.h>.

Not yet fully standalone: some dependency headers (e.g. tcMesh.h) are still
order-dependent on the umbrella. Making the header tree self-contained is a
separate, ongoing effort; this is a step toward it.
